pyproject.toml
[project]
name = "multiplatform-deep-researcher"
version = "0.1.0"
description = "Implementing deep research using CrewAI and brightdata web mcp"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Core Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "gunicorn>=23.0.0",
    
    # CrewAI & MCP
    "crewai-tools[mcp]>=0.69.0",
    "mcp>=1.13.1",
    
    # Rate Limiting
    "slowapi>=0.1.9",
    
    # Utilities
    "python-dotenv>=1.1.1",
    "httpx>=0.28.0",
    
    # Optional: Streamlit UI
    "streamlit>=1.49.1",
    "ollama>=0.5.3",
]
